Add package doc and move LongAESRF to aes package
