+++ src/main/java/org/cryptimeleon/math/prf/aes/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains an implementation of an AES-based pseudorandom function.
  */
-package org.cryptimeleon.craco.prf.aes;
+package org.cryptimeleon.math.prf.aes;
