+++ src/main/java/org/cryptimeleon/math/prf/zn/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains an implementation of a Hash-then-Prf-to-Zn construction.
+ */
+package org.cryptimeleon.math.prf.zn;
