+++ src/test/java/org/cryptimeleon/math/serialization/standalone/params/PrfStandaloneReprTests.java
@@ -3,7 +3,7 @@ package org.cryptimeleon.math.serialization.standalone.params;
 import org.cryptimeleon.math.hash.impl.SHA256HashFunction;
 import org.cryptimeleon.math.prf.aes.AesPseudorandomFunction;
 import org.cryptimeleon.math.prf.zn.HashThenPrfToZn;
-import org.cryptimeleon.math.prf.zn.LongAesPseudoRandomFunction;
+import org.cryptimeleon.math.prf.aes.LongAesPseudoRandomFunction;
 import org.cryptimeleon.math.serialization.standalone.StandaloneReprSubTest;
 import org.cryptimeleon.math.structures.rings.zn.Zn;
